Refactoring

+++ src/main/java/ch/hsr/zebrastreifensafari/jpa/entities/Crossing.java
@@ -47,7 +47,7 @@ public class Crossing implements Serializable {
         this.id = id;
     }
 
-    public Crossing(Integer id, long osmNodeId, int ratingAmount, int status) {
+    public Crossing(Integer id, long osmNodeId, long ratingAmount, int status) {
         this.id = id;
         this.osmNodeId = osmNodeId;
         this.ratingAmount = ratingAmount;
